Refactor : Moved queue into different package
