+++ utils/helper.go
@@ -17,7 +17,13 @@ var ExponentialBackoffRetry = func(maxRetries int, operation func() error) error
 		}
 		logrus.Errorf("Attempt %d: Operation failed: %s", i+1, err)
 		if i < maxRetries-1 {
-			time.Sleep(time.Duration(math.Pow(2, float64(i))) * time.Second)
+			backoffDuration := time.Duration(math.Pow(2, float64(i))) * time.Second
+
+			if backoffDuration > 30*time.Second {
+				backoffDuration = 30 * time.Second
+			}
+			time.Sleep(backoffDuration)
+
 		}
 	}
 	return err
